answer (metric value) view/method reorganization
