+++ mod/metric_answer_lookup/lib/answer.rb
@@ -15,6 +15,11 @@ class Answer < ApplicationRecord
   validate :must_be_an_answer, :card_must_exist, unless: :virtual?
   validate :metric_must_exit
 
+  def self.existing id
+    return unless id
+    find_by_answer_id(id) || (refresh(id) && find_by_answer_id(id))
+  end
+
   def card_column
     :answer_id
   end
