+++ mod/metrics/set/abstract/expanded_researched_details.rb
@@ -1,5 +1,5 @@
 format :html do
-  def wrap_value_details
+  def wrap_expanded_details
     output [
       _render(:credit_name),
       yield,
@@ -7,10 +7,10 @@ format :html do
     ]
   end
 
-  view :researched_value_details do
+  view :expanded_researched_details do
     checked_by = card.fetch trait: :checked_by, new: {}
     checked_by = nest(checked_by, view: :core)
-    wrap_value_details do
+    wrap_expanded_details do
       [
         wrap_with(:div, checked_by, class: "double-check mt-3"),
         wrap_with(:div, _render_sources, class: "cited-sources mt-3")
