+++ mod/metrics/set/type/metric_value.rb
@@ -1,6 +1,11 @@
 include_set Abstract::MetricChild, generation: 2
 include_set Abstract::MetricAnswer
 
+def answer
+  @answer ||= Answer.existing(id) || Answer.new
+end
+
+# EVENTS
 event :flash_success_message, :finalize, on: :create do
   msg =
     format(:html).alert :success, true, false, class: "text-center" do
@@ -11,37 +16,16 @@ event :flash_success_message, :finalize, on: :create do
   success.flash msg
 end
 
-def filtered_item_query filter={}, sort={}, paging={}
-  filter[:year] = year.to_i
-  FixedMetricAnswerQuery.new metric_card.id, filter, sort, paging
-end
-
-def answer
-  @answer ||= existing_answer || Answer.new
-end
-
-def existing_answer
-  return unless id
-  Answer.find_by_answer_id(id) ||
-    (Answer.refresh(id) && Answer.find_by_answer_id(id))
-end
-
-format :json do
-  view :core do
-    data = _render_essentials.merge(
-      metric: nest(card.metric, view: :essentials),
-      company: nest(card.company, view: :marks)
-    )
-    data[:source] = nest(card.source, view: :essentials) if card.source.present?
-    data.merge(checked_by: nest(card.checked_by_card, view: :essentials, hide: :marks))
+# AS RESEARCH PAGE
+format :html do
+  view :open_content, cache: :never do
+    voo.hide! :cited_source_links
+    subformat(:research_page).slot_machine metric: card.metric, company: card.company,
+                                           year: card.year # active_tab: "View Source"
   end
 
-  def essentials
-    {
-      year: card.year,
-      value: card.value,
-      import: card.imported?,
-      comments: field_nest(:discussion, view: :core)
-    }
+  def default_title_args _args
+    # HACK: to prevent cancel button on research page from loosing title
+    voo.title ||= "Answer"
   end
 end
