+++ mod/research/set/self/research_page/_metric_slot.haml
@@ -31,7 +31,7 @@
                     %div.mr-3
                       = a.year
                     %div
-                      = nest a, view: :conciser, hide: :year
+                      = nest a, view: :conciser, hide: :year_equals
             .dropdown-divider
   - if project?
     .d-flex.flex-nowrap.align-items-center.ml-3.w-25.justify-content-between
